% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmrmod.R
\name{mmrmod}
\alias{mmrmod}
\title{MMRM analysis without assuming orthogonality property}
\usage{
mmrmod(
  data,
  outcome,
  group,
  time,
  subject,
  covariate = NULL,
  covfactor = NULL,
  type = "UN",
  hetero = TRUE,
  robust = TRUE,
  ssadjust = 1,
  reml = FALSE,
  intcov = TRUE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{a data frame that may include \code{outcome}, \code{group},
\code{time}, \code{id}, and specified covariate variables.}

\item{outcome}{a name of outcome (dependent) variable included in
\code{data}.}

\item{group}{a name of treatment group variable included in \code{data}.}

\item{time}{a name of time variable for repeated measurements included
in \code{data}. The default is \code{NULL}.}

\item{subject}{a name of subject id variable for repeated measurements
included in \code{data}. The default is \code{NULL}.}

\item{covariate}{a character vector with names of covariate variables
included in \code{data}. The default is \code{NULL}.}

\item{covfactor}{an integer vector including nominal variable indicators for
covariate variables. Nominal variable: \code{1}, continuous variable:
\code{0}. The default is \code{NULL}.}

\item{type}{specify the covariance structure type from \code{c("UN", "CS",
"AR(1)")}. The default is \code{"UN"}.}

\item{hetero}{a logical value specifying whether or not to use a unequal
variance model among groups.
Unequal variance: \code{TRUE}, equal variance: \code{FALSE}.
Default is \code{TRUE}.}

\item{robust}{a logical value specifying whether or not to use robust
inference with sandwich-type standard error.
Robust inference: \code{TRUE}, naive inference: \code{FALSE}.
Default is \code{TRUE}.}

\item{ssadjust}{an integer specifying whether small sample adjustment is
performed. No adjustment: \code{0}, adjustment method 1: \code{1},
adjustment method 2: \code{2}. For details on each method, see
Maruo et al. (submitting). Default is \code{1}.}

\item{reml}{a logical value specifying whether to use the REML inference.
If orthogonality is not assumed, the REML inference cannot be used, so
if \code{REML=TRUE} is specified, only an \code{mmrm} object by
the \code{mmrm} package is returned. Default is \code{FALSE}.}

\item{intcov}{a logical value specifying whether to include interactions
between \code{time} and \code{covariate}. Default is \code{TRUE}.}

\item{conf.level}{a numeric value of the confidence level for the
confidence intervals. Default is 0.95.}

\item{...}{additional arguments to be passed to the \code{mmrm} function.
For example, it can identify an optimizer in parameter estimation.
Note that the Kenward-Roger method cannot be used in degree-of-freedom
estimation unless \code{REML=TRUE} is specified.
See \code{\link{mmrm}} function for details.}
}
\value{
a list that includes following components
for the MMRM analysis:
\describe{
\item{\code{marg.mean}}{a data frame including inference results of
marginal mean for each level of \code{time} and \code{group}.}
\item{\code{diff.mean}}{a data frame including inference results of
difference of marginal mean between \code{group} at each \code{time}.
}
\item{\code{Vtheta}}{a variance-covariance matrix for MLE of
the vector of all parameters.}
\item{\code{mmrmObject}}{an "\code{mmrm}" object by \code{mmrm} package. }
}
}
\description{
\code{mmrmod} provides inference results for the mixed models for repeated
measures (MMRM) analysis without assuming orthogonality property between
fixed effect and variance-covariance parameters.
}
\details{
The inference without assuming orthogonality property means that the
off-diagonal ("od") blocks of the variance-covariance matrix of the
parameters, i.e., the covariance between the fixed effects parameter and
the covariance parameter, are not zero.

The usual linear mixed model program packages make inferences on the
assumption that the off-diagonal block is a zero matrix from the properties
of the normal distribution, however, this assumption is incorrect when
the normality of the error does not hold and/or the missing values are not
from MCAR mechanism.

The function uses the \code{mmrm} function from the \code{mmrm} package
for parameter estimation.
}
\examples{
library(mmrm)
data(fev_data)

mmrmod(data = fev_data, outcome = FEV1, group = ARMCD, time = AVISIT,
       subject = USUBJID, covariate = c("RACE", "SEX", "FEV1_BL"),
       covfactor = c(1, 1, 0))

}
\references{
\itemize{
\item Maruo K, Ishii R, Yamaguchi Y, Doi M, Gosho M (2020). A note on the
bias of standard errors when orthogonality of mean and variance
parameters is not satisfied in the mixed model for repeated measures
analysis. Statistics in Medicine, 39, 1264-1274,
\url{https://doi.org/10.1002/sim.8474}.\cr
\item Maruo K, Ishii R, Yamaguchi Y, Gosho M (submitting). Small sample
adjustment for inference without assuming orthogonality in MMRM
analysis.
\item Sabanes Bove D, Dedic J, Kelkhoff D, Kunzmann K, Lang B, Li L, Wang Y
(2022). mmrm: Mixed models for repeated measures. R package version
0.2.2, \url{https://CRAN.R-project.org/package=mmrm}.
}
}
\seealso{
\code{\link{mmrm}}
}
